refactor: refactor the code in modules.self_attention and rename the package into "transformer";
